refactoring: move custom_filter.rs.
