+++ util/indexer-sync/Cargo.toml
@@ -14,10 +14,14 @@ repository = "https://github.com/nervosnetwork/ckb"
 ckb-app-config = { path = "../app-config", version = "= 0.113.0-pre" }
 ckb-async-runtime = { path = "../runtime", version = "= 0.113.0-pre" }
 ckb-db-schema = { path = "../../db-schema", version = "= 0.113.0-pre" }
+ckb-jsonrpc-types = { path = "../jsonrpc-types", version = "= 0.113.0-pre" }
 ckb-logger = { path = "../logger", version = "= 0.113.0-pre" }
 ckb-notify = { path = "../../notify", version = "= 0.113.0-pre" }
 ckb-stop-handler = { path = "../stop-handler", version = "= 0.113.0-pre" }
 ckb-store = { path = "../../store", version = "= 0.113.0-pre" }
 ckb-types = { path = "../types", version = "= 0.113.0-pre" }
+numext-fixed-uint = "0.1"
+rhai = { version = "1.10.0", features = ["no_function", "no_float", "no_module", "sync"]}
 rocksdb = { package = "ckb-rocksdb", version ="=0.21.1", features = ["snappy"], default-features = false }
+serde_json = "1.0"
 thiserror = "1.0"
