+++ util/indexer/Cargo.toml
@@ -17,8 +17,6 @@ ckb-indexer-sync = { path = "../indexer-sync", version = "= 0.113.0-pre" }
 ckb-jsonrpc-types = { path = "../jsonrpc-types", version = "= 0.113.0-pre" }
 ckb-notify = { path = "../../notify", version = "= 0.113.0-pre" }
 ckb-types = { path = "../types", version = "= 0.113.0-pre" }
-numext-fixed-uint = "0.1"
-rhai = { version = "1.10.0", features = ["no_function", "no_float", "no_module", "sync"]}
 rocksdb = { package = "ckb-rocksdb", version ="=0.21.1", features = ["snappy"], default-features = false }
 serde_json = "1.0"
 thiserror = "1.0"
